+++ c/src/sort/quick_sort.c
@@ -1,11 +1,11 @@
 #include "sort.h"
 
-unsigned int partition(int* arr, unsigned int size) {
+size_t partition(int* arr, size_t size) {
   int aux = 0;
-  unsigned int last_pos = size - 1;
+  size_t last_pos = size - 1;
   int pivot = arr[last_pos];
-  unsigned int pivot_position = 0;
-  for (unsigned int i = 0; i < size; ++i) {
+  size_t pivot_position = 0;
+  for (size_t i = 0; i < size; ++i) {
     if (arr[i] < pivot) {
       aux = arr[pivot_position];
       arr[pivot_position++] = arr[i];
@@ -18,17 +18,17 @@ unsigned int partition(int* arr, unsigned int size) {
   return pivot_position;
 }
 
-void quick_sort(int* arr, unsigned int size) {
+void quick_sort(int* arr, size_t size) {
   if (size > 1) {
-    unsigned int pivot_position = partition(arr, size);
+    size_t pivot_position = partition(arr, size);
     quick_sort(arr, pivot_position);
     quick_sort(arr + pivot_position + 1, size - pivot_position - 1);
   }
 }
 
-unsigned int partition_ram(int* arr, unsigned int size,
+size_t partition_ram(int* arr, size_t size,
                            unsigned long long* op) {
   return 0;
 }
 
-unsigned long long quick_sort_ram(int* arr, unsigned int size) { return 0; }
+unsigned long long quick_sort_ram(int* arr, size_t size) { return 0; }
