+++ c/src/sort/randomized_quick_sort.c
@@ -2,15 +2,15 @@
 
 #include "sort.h"
 
-unsigned int randomized_partition(int* arr, unsigned int size) {
-  unsigned int last_pos = size - 1;
-  unsigned int random_pos = rand() % size;
+size_t randomized_partition(int* arr, size_t size) {
+  size_t last_pos = size - 1;
+  size_t random_pos = rand() % size;
   int pivot = arr[random_pos];
   arr[random_pos] = arr[last_pos];
   arr[last_pos] = pivot;
-  unsigned int pivot_position = 0;
+  size_t pivot_position = 0;
   int aux;
-  for (unsigned int i = 0; i < size; ++i) {
+  for (size_t i = 0; i < size; ++i) {
     if (arr[i] < pivot) {
       aux = arr[pivot_position];
       arr[pivot_position++] = arr[i];
@@ -23,19 +23,19 @@ unsigned int randomized_partition(int* arr, unsigned int size) {
   return pivot_position;
 }
 
-void randomized_quick_sort(int* arr, unsigned int size) {
+void randomized_quick_sort(int* arr, size_t size) {
   if (size > 1) {
-    unsigned int pivot_position = randomized_partition(arr, size);
+    size_t pivot_position = randomized_partition(arr, size);
     randomized_quick_sort(arr, pivot_position);
     randomized_quick_sort(arr + pivot_position + 1, size - pivot_position - 1);
   }
 }
 
-unsigned int randomized_partition_ram(int* arr, unsigned int size,
+size_t randomized_partition_ram(int* arr, size_t size,
                                       unsigned long long* op) {
   return 0;
 }
 
-unsigned long long randomized_quick_sort_ram(int* arr, unsigned int size) {
+unsigned long long randomized_quick_sort_ram(int* arr, size_t size) {
   return 0;
 }
